Fix Windows build (#175)
* wip

* Fix visibility in rviz_common

- needs to export some more symbols from rviz_common
- small uncrustify changes

* Fix stl_loader

* Use dynamically linked OgrePlugins

Don't link against Plugins anymore (loaded dynamically)

Set ogre_plugin_dir to load dynamically

Release and Debug need different behaviour

* Fix yaml-cpp dynamic linking

see https://github.com/jbeder/yaml-cpp/issues/332

* Install rviz_rendering dlls

* Reenable Windows tests

* Move test folder to correct place

- this is now consistent with rviz_common and rviz_default_plugins
- fixup location of ogre_testing_environment
- ogre_testing_environment needs to be exported

* Fix linking against Ogre

* Add zlib to be build on Windows

* Change targets to unknown instead of shared

* Fix billboard_line tests for Windows

- need to expose symbols of billboard line for Windows
- copy constructor of billboard line is not generated on Windows

* Use rviz_common::UniformStringStream instead of std::stringstream

* Fix linking of tests on Windows

- just like rviz2.exe, tests need dll directories in %PATH%
- add local install to environment variables for testing
- in isolated builds, ogre resources cannot be found otherwise
- enable tests in rviz_default_plugins on Windows but disable display

* Fix grid_test

* Linking to exported template function is problematic on Windows

- this does not work with RVIZ_DEFAULT_PLUGINS_PUBLIC
- delete tests for templated function
- hide templated function again

* Fix image_display tests

* Introduce visibility control in rviz_default_plugins

* run windeployqt on rviz2 binary to make it runnable after installation

* extend PATH on Windows with location of libraries for vendor packages

* Add missing visibility_control.hpp and fix license

* log build output of dependencies to file

* Fix Windows warnings

* delete visibility from rviz_default_plugins

* Disable deprecation warnings in Ogre

* Fix dllexport warning in ros abstraction

* fix linking warnings in rviz_rendering

* fix linking errors

* Fix warning in rviz_rendering

* Reintroduce destructor to fix test

+++ rviz_common/src/rviz_common/selection/selection_manager.cpp
@@ -759,7 +759,7 @@ bool SelectionManager::render(
 
   Ogre::PixelFormat format = pixel_buffer->getFormat();
 
-  int size = Ogre::PixelUtil::getMemorySize(render_w, render_h, 1, format);
+  int size = static_cast<int>(Ogre::PixelUtil::getMemorySize(render_w, render_h, 1, format));
   uint8_t * data = new uint8_t[size];
 
   delete[] reinterpret_cast<uint8_t *>(dst_box.data);
